Hack to allow date columns to be set to null
See https://github.com/elixir-ecto/tds/issues/124 for more details. This
fix is a complete hack that gets the job done for date columns but fails
for the same reason on binary columns. Thankfully for us, we don't have
any binary columns, so this works until a correct fix can be made.